+++ src/Navigation.php
@@ -54,7 +54,7 @@ class Navigation extends Plugin
 
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '2.0.1';
+    public string $schemaVersion = '2.0.2';
     public string $minVersionRequired = '1.4.24';