+++ src/controllers/NodesController.php
@@ -105,12 +105,6 @@ class NodesController extends Controller
 
         $node->newParentId = $request->getParam("{$prefix}parentId");
 
-        // Handle custom URL - remove the elementId. Particularly if we're swapping
-        if ($node->isManual()) {
-            $node->elementId = null;
-            $node->elementSiteId = null;
-        }
-
         // Set field values.
         $node->setFieldValuesFromRequest('fields');