+++ src/nodetypes/CustomType.php
@@ -5,14 +5,14 @@ use verbb\navigation\base\NodeType;
 
 use Craft;
 
-class PassiveType extends NodeType
+class CustomType extends NodeType
 {
     // Static Methods
     // =========================================================================
 
     public static function displayName(): string
     {
-        return Craft::t('navigation', 'Passive');
+        return Craft::t('navigation', 'Custom URL');
     }
 
     public static function hasTitle(): bool
@@ -22,12 +22,12 @@ class PassiveType extends NodeType
 
     public static function hasUrl(): bool
     {
-        return false;
+        return true;
     }
 
     public static function hasNewWindow(): bool
     {
-        return false;
+        return true;
     }
 
     public static function hasClasses(): bool
@@ -37,6 +37,17 @@ class PassiveType extends NodeType
 
     public static function hasAttributes(): bool
     {
-        return true;
+        return false;
+    }
+
+
+    // Public Methods
+    // =========================================================================
+
+    public function getModalHtml(): ?string
+    {
+        return Craft::$app->getView()->renderTemplate('navigation/_types/custom/modal', [
+            'node' => $this->node,
+        ]);
     }
 }